+++ packages/table/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.5.9 (2022-02-21)
+
+**Note:** Version bump only for package @equinor/fusion-react-table
+
+
+
+
+
 ## [0.5.8](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-table@0.5.7...@equinor/fusion-react-table@0.5.8) (2022-02-11)
 
 **Note:** Version bump only for package @equinor/fusion-react-table